fix: error message is incorrect when a host is not in AWS
This fixes two issues. The first is that when a host wasn't found, the
page would show the red error text because the `input` variable was no
longer defined; this is supposed to be the new `lookup` variable
instead. This restores the correct function of the blue "not found"
error instead of always showing an unclear red error.

The second is that saying a host name isn't within AWS IP space might
be confusing so instead the wording now says "hosted" within.